format change

+++ bookkeeping_website/iot/models.py
@@ -7,7 +7,7 @@ class Event(models.Model):
     node_loc = models.CharField(max_length=10)
     temp = models.DecimalField(max_digits=5, decimal_places=2)
     humid = models.DecimalField(max_digits=5, decimal_places=2)
-    light = models.IntegerField()
+    light = models.DecimalField(max_digits=5, decimal_places=2)
     sound = models.DecimalField(max_digits=5, decimal_places=2)
     date_created = models.DateTimeField(auto_now_add=True)
     def __str__(self):
